Projects/ai-content-studio/requirements.txt
# AI Content Studio - Requirements

# Core Framework
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# AI/LLM Integration
anthropic>=0.25.0
openai>=1.0.0

# Voice Synthesis
elevenlabs>=0.2.27

# Image/Video Processing
opencv-python>=4.8.1
pillow>=10.1.0
moviepy>=1.0.3

# Audio Processing
pydub>=0.25.1
ffmpeg-python>=0.2.0

# Storage & Queue
boto3>=1.34.0
redis>=5.0.1

# Web Framework (for future upload portal)
fastapi>=0.109.0
uvicorn>=0.25.0
python-multipart>=0.0.6

# Utilities
aiohttp>=3.9.1
requests>=2.31.0
jinja2>=3.1.2
pyyaml>=6.0.1
python-slugify>=8.0.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development
black>=23.12.0
isort>=5.13.2
mypy>=1.7.1
